+++ lib/utils/component_utils.dart
@@ -180,4 +180,13 @@ class ComponentUtils {
       ),
     );
   }
+
+  static void showSnackbar(BuildContext context, String message, {bool isError = false}) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: isError ? Colors.red : Colors.green,
+      ),
+    );
+  }
 }
