+++ lib/views/concrete_quality_search.dart
@@ -51,6 +51,7 @@ class _ConcreteQualitySearchState extends State<ConcreteQualitySearch> {
         child: Padding(
           padding: const EdgeInsets.all(16.0),
           child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               CustomDropdownFormField(
                 labelText: "Registro",
