+++ pubspec.yaml
@@ -7,10 +7,11 @@ environment:
   sdk: '>=3.2.6 <4.0.0'
 
 dependencies:
+  data_table_2: ^2.5.15
   flutter:
     sdk: flutter
   flutter_dotenv: ^5.1.0
-  fluttertoast: ^8.2.5
+  fluttertoast: ^8.2.6
   injector: ^3.0.0
   intl: ^0.19.0
   path: ^1.8.3
